// frontend/src/types/Supporter.ts
export interface Supporter {
  supporterId: number;
  supporterType?: string;
  displayName?: string;
  organizationName?: string;
  firstName?: string;
  lastName?: string;
  relationshipType?: string;
  region?: string;
  country?: string;
  email?: string;
  phone?: string;
  status?: string;
  createdAt?: string;
  firstDonationDate?: string;
  acquisitionChannel?: string;
}